requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
pytest==7.4.0